requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
requests==2.32.3
beautifulsoup4==4.12.3
dnspython==2.6.1
shodan==1.31.0
boto3==1.34.135
python-dotenv==1.0.1
reportlab==4.2.2
pandas==2.2.2
python-multipart==0.0.9
aiofiles==23.2.1
